Create handle of demo app

+++ src/app/useCases/ingredients/deleteIngredient.ts
@@ -6,6 +6,8 @@ export async function deleteIngredient(req: Request, res: Response) {
   try {
     const { ingredientId } = req.params;
 
+    if (req.headers['demo'] === 'true') return res.sendStatus(204);
+
     await Ingredient.findByIdAndDelete(ingredientId);
 
     res.sendStatus(204);
