+++ src/app/useCases/orders/cancelOrder.ts
@@ -6,6 +6,8 @@ export async function cancelOrder(req: Request, res: Response) {
   try {
     const { orderId } = req.params;
 
+    if (req.headers['demo'] === 'true') return res.sendStatus(204);
+
     await Order.findByIdAndDelete(orderId);
 
     res.sendStatus(204);
