+++ src/app/useCases/users/deleteUser.ts
@@ -6,6 +6,8 @@ export async function deleteUser(req: Request, res: Response) {
   try {
     const { userId } = req.params;
 
+    if (req.headers['demo'] === 'true') return res.sendStatus(204);
+
     await User.findByIdAndDelete(userId);
 
     res.sendStatus(204);
